+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+- Don't fail when building the manual with AsciiDoc and xmllint.
+
 ## [1.1.1] - 2018-11-07
 
 ### Fixed
@@ -23,6 +28,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 - The initial release, which evolved from an early prototype specific to a
   certain use case into a more generic command line tool.
 
+[Unreleased]: https://github.com/nixcloud/ip2unix/compare/v1.1.1...HEAD
 [1.1.1]: https://github.com/nixcloud/ip2unix/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/nixcloud/ip2unix/compare/v1.0.0...v1.1.0