+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ### Fixed
 - Don't fail when building the manual with AsciiDoc and xmllint.
+- Some long options (`--rule`, `--rules-file` and `--rules-data`) were ignored.
 
 ### Added
 - Create a release.nix for building and testing with Hydra.