+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ### Added
 - Create a release.nix for building and testing with Hydra.
+- A new --version command line argument for showing version information.
 
 ## [1.1.1] - 2018-11-07