+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
-## [Unreleased]
+## [1.2.0] - 2018-11-12
 
 ### Fixed
 - Don't fail when building the manual with AsciiDoc and xmllint.
@@ -33,7 +33,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 - The initial release, which evolved from an early prototype specific to a
   certain use case into a more generic command line tool.
 
-[Unreleased]: https://github.com/nixcloud/ip2unix/compare/v1.1.1...HEAD
+[1.2.0]: https://github.com/nixcloud/ip2unix/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/nixcloud/ip2unix/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/nixcloud/ip2unix/compare/v1.0.0...v1.1.0